refactor(solana): change internal structure of @trezor/coins-solana
