+++ coins/coins-solana/src/runtime/exports.ts
@@ -1,7 +1,7 @@
 export * from './api';
-export * from './common';
 export * from './connect';
 export * from './fees';
 export * from './signing';
 export * from './staking';
-export * from './transactions';
+export * from './shim';
+export * from './utils';
