+++ suite-common/wallet-config/src/stakingProviders.ts
@@ -1,8 +1,5 @@
 import { EVERSTAKE_POOLS, FIVE_BINARIES_POOLS } from '@suite-common/wallet-constants';
-import {
-    EVERSTAKE_SOLANA_DEVNET_VALIDATOR,
-    EVERSTAKE_SOLANA_MAINNET_VALIDATOR,
-} from '@trezor/coins-solana/constants';
+import { EVERSTAKE_VOTER_PUBKEYS } from '@trezor/coins-solana/constants';
 
 export type StakingProviderId = 'everstake' | 'fivebinaries';
 
@@ -17,7 +14,7 @@ export type StakingProvider = {
 export const EVERSTAKE_PROVIDER: StakingProvider = {
     id: 'everstake',
     name: 'Everstake',
-    solanaVoterPubkeys: [EVERSTAKE_SOLANA_MAINNET_VALIDATOR, EVERSTAKE_SOLANA_DEVNET_VALIDATOR],
+    solanaVoterPubkeys: EVERSTAKE_VOTER_PUBKEYS,
     cardanoPoolIds: EVERSTAKE_POOLS,
     ethereumPoolNames: ['Everstake'],
 };
